add task to generate ViewHolderProvider
