+++ linen-starter/src/main/java/x7c1/linen/MainActivity.java
@@ -19,8 +19,8 @@ import android.widget.TextView;
 
 import x7c1.linen.modern.SampleAdapter;
 import x7c1.linen.modern.SampleImpl;
-import x7c1.linen.res.SourceRowProvider;
 import x7c1.linen.res.layout.CommentRowLayoutProvider;
+import x7c1.linen.res.layout.SourceRowProvider;
 import x7c1.linen.res.values.CommentValuesProvider;
 
 import static x7c1.linen.FlingDetector.forHorizontal;
