+++ linen-starter/src/main/java/x7c1/linen/res/layout/SourceRowProvider.java
@@ -1,0 +1,40 @@
+
+/**
+ * This file is automatically generated by wheat-build.
+ * Do not modify this file -- YOUR CHANGES WILL BE ERASED!
+ */
+
+package x7c1.linen.res.layout;
+
+import android.content.Context;
+import android.view.LayoutInflater;
+import android.view.ViewGroup;
+import android.view.View;
+import android.widget.TextView;
+import x7c1.wheat.ancient.resource.ViewHolderProvider;
+import x7c1.linen.R;
+import x7c1.linen.glue.res.layout.SourceRow;
+
+public class SourceRowProvider implements ViewHolderProvider<SourceRow> {
+
+    private final LayoutInflater inflater;
+
+    public SourceRowProvider(Context context){
+        inflater = LayoutInflater.from(context);
+    }
+
+    @Override
+    public SourceRow inflateOn(ViewGroup parent){
+        return inflate(parent, false);
+    }
+
+    @Override
+    public SourceRow inflate(ViewGroup parent, boolean attachToRoot){
+        View view = inflater.inflate(R.layout.source_row, parent, attachToRoot);
+        return new SourceRow(
+            view,
+            (TextView) view.findViewById(R.id.source_row__title),
+            (TextView) view.findViewById(R.id.source_row__description)
+        );
+    }
+}
